realocandoas funções criadas nas etapas de EDA e data scraping na pasta notebooks/functions para a pasta src
